# Optimized height calculation for map preview