--- a/src/map_io/widelands_map_loader.cc
+++ b/src/map_io/widelands_map_loader.cc
@@ -172,7 +172,7 @@ int32_t WidelandsMapLoader::load_map_for_render(EditorGameBase& egbase, AddOns::
 		p.read(*fs_, egbase, false, *mol_);
 	}
 
-	map_.recalc_whole_map(egbase);
+	map_.recalc_whole_map_brightness();
 
 	return 0;
 }